FIX: in pandas 0.12.0 df.index.values returns an array of dtype object which caused error in np.assert_allclose; cast explicitly to np.float to resolve this issue.

+++ seaborn/tests/test_timeseries_new.py
@@ -854,7 +854,10 @@ class TestPlotsNew(PlotTestCase):
             gammas_c = self.gammas[idx_c]
             gammas_c_gb = gammas_c.groupby(self.gammas_kwargs['time'])
             y = gammas_c_gb[self.gammas_kwargs['value']].agg(np.mean)
-            npt.assert_allclose(line.get_xdata(), y.index.values)
+            # .astype(np.float) because in pandas 0.12.0
+            # y.index.values.dtype is of dtype 'object'
+            npt.assert_allclose(line.get_xdata(),
+                                y.index.values.astype(np.float))
             npt.assert_allclose(line.get_ydata(), y.values)
 
 
@@ -1106,5 +1109,8 @@ class TestPlotsOld(PlotTestCase):
             gammas_c = self.gammas[idx_c]
             gammas_c_gb = gammas_c.groupby(self.gammas_kwargs['time'])
             y = gammas_c_gb[self.gammas_kwargs['value']].agg(np.mean)
-            npt.assert_allclose(line.get_xdata(), y.index.values)
+            # .astype(np.float) because in pandas 0.12.0
+            # y.index.values.dtype is of dtype 'object'
+            npt.assert_allclose(line.get_xdata(),
+                                y.index.values.astype(np.float))
             npt.assert_allclose(line.get_ydata(), y.values)
